Seperating logic
